+++ src/includes/wp-sharks-core-rv.php
@@ -9,7 +9,7 @@ declare (strict_types = 1);
 namespace WebSharks\WpSharks\WooCommerce\Restrictions\Pro;
 
 if (!defined('WPINC')) {
-    exit('Do NOT access this file directly: '.basename(__FILE__));
+    exit('Do NOT access this file directly.');
 }
 $GLOBALS['wp_sharks_core_rv'] = [
     'min' => '160828.35574', //wp-sharks-core-required-version//